% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineCoxreg2.R
\name{splineCox.reg2}
\alias{splineCox.reg2}
\title{Fitting the five-parameter spline Cox model with a specified shape, selecting the best fit}
\usage{
splineCox.reg2(
  t.event,
  event,
  Z,
  xi1 = min(t.event),
  xi3 = max(t.event),
  model = names(shape.list),
  p0 = rep(0, 1 + ncol(as.matrix(Z)))
)
}
\arguments{
\item{t.event}{a vector for time-to-event}

\item{event}{a vector for event indicator (=1 event; =0 censoring)}

\item{Z}{a matrix for covariates; nrow(Z)=sample size, ncol(Z)=the number of covariates}

\item{xi1}{lower bound for the hazard function; the default is min(t.event)}

\item{xi3}{upper bound for the hazard function; the default is max(t.event)}

\item{model}{A character vector specifying which model shapes to consider for the baseline hazard.
Available options are:
"increase", "constant", "decrease", "unimodal1", "unimodal2", "unimodal3", "bathtub1", "bathtub2", "bathtub3".
Default is \code{names(shape.list)} which includes all available models.}

\item{p0}{Initial values to maximize the likelihood (1 + p parameters; baseline hazard scale parameter and p regression coefficients)}
}
\value{
A list containing the following components:
\item{model}{A character string indicating the shape of the baseline hazard function used.}
\item{parameter}{A numeric vector of the parameters defining the baseline hazard shape.}
\item{beta}{A named vector with the estimates, standard errors, and 95\% confidence intervals for the regression coefficients}
\item{gamma}{A named vector with the estimate, standard error, and 95\% confidence interval for the baseline hazard parameter}
\item{loglik}{A named vector containing the log-likelihood (\code{LogLikelihood}),
Akaike Information Criterion (\code{AIC}), and Bayesian Information
Criterion (\code{BIC}) for the best-fitting model}
\item{other_models}{A data frame containing the log-likelihood (\code{LogLikelihood}) for all other evaluated models,
with model names as row names.}
}
\description{
\code{splineCox.reg2} estimates the parameters of a five-parameter spline Cox model for multiple specified shapes
and selects the best fitting model based on the minimization of the log-likelihood function.
The function calculates the estimates for the model parameters (beta) and the baseline hazard scale parameter (gamma), using non-linear optimization.
}
\examples{
# Example data
library(joint.Cox)
data(dataOvarian)
t.event = dataOvarian$t.event
event = dataOvarian$event
Z = dataOvarian$CXCL12

M = c("constant", "increase", "decrease")
reg2 <- splineCox.reg2(t.event, event, Z, model = M)
print(reg2)

}
